test: Move `stdlib::time` tests to integration file and fix race condition
Following PR feedback, the inline tests originally added to `src/stdlib/time.rs`
have been removed and appended to the existing integration test file
`tests/time_stdlib_test.rs`. The code was correctly formatted using `cargo fmt`,
and a midnight-rollover race condition in `test_native_today` was resolved by
capturing the expected date before calling the function under test.
